+++ src/main/kotlin/app/revanced/patches/messenger/inputfield/fingerprints/SwitchMessangeInputEmojiButtonFingerprint.kt
@@ -1,0 +1,18 @@
+package app.revanced.patches.messenger.inputfield.fingerprints
+
+import app.revanced.patcher.fingerprint.method.impl.MethodFingerprint
+import org.jf.dexlib2.Opcode
+
+object SwitchMessangeInputEmojiButtonFingerprint : MethodFingerprint(
+    returnType = "V",
+    parameters = listOf("L", "Z"),
+    strings = listOf("afterTextChanged", "expression_search"),
+    opcodes = listOf(
+        Opcode.IGET_OBJECT,
+        Opcode.IF_EQZ,
+        Opcode.CONST_STRING,
+        Opcode.GOTO,
+        Opcode.CONST_STRING,
+        Opcode.GOTO
+    )
+)